treat prototype objects as plain objects
The prototype object associated with some type T is not a member of T,
even if its '@@type' property indicates that it is. The object may be
missing fields set by T's data constructors. The Left and Right data
constructors of the Either type, for example, could define 'isLeft',
'isRight', and 'value' fields which various methods assume to exist.
The following method invocation is therefore incorrect:

    Either.prototype['fantasy-land/equals'](x)

This commit adds `x.constructor.prototype !== x` checks to `type` and
`getBoundMethod` so that prototype objects are treated as plain objects.

'@@type' properties should now be defined on type representatives rather
than on prototype objects. This is a breaking change.