Fix incorrect usage of arb.smap in tests
As I understand, the .smap method is not a map function. Its intended use
is to provide a way for JSVerify to verify that, given a set, a reduction
of that set through a given function will produce a succesfull assertion
if the greater set fails the same assertion.

It takes two functions: f, and g, where g should be the right inverse of
f. The current code only passes f, and leaves g to undefined. This means
that if JSVerify would ever find a failing property using one of these
incorrectly smapped arbitraries, and deem it necessary to attempt
shrinking it, it would throw an error like: 'g is not a function'.

This commit makes two changes:

1. A 'smapArb' dispatcher was added which forces both arguments to be
   provided. All occurrences of the arb.smap call where a right inverse
   function could be defined have been replaced to use this new
   dispatcher.
2. A 'mapArb' dispatcher was added which can be used safely in the way
   that arb.smap was previously used. It relies on jsc.bless to attach a
   default shrink function. The default shrink function just produces an
   empty set.