Add tests to assert explicitly that Z copes with alien constructors
By alien constructor, I mean values that carry around constructors from
a different global scope that the one being used by the currently
running code. Sanctuary Type Classes has always supported this by
relying on Sanctuary Type Identifiers to return string representations
of the constructors that can be compared reliably.

In the following commits, refactors risk removing this feature because
it is not captured in the unit tests. This commit should ensure that the
feature is not lost.