Improve performance of equals and lte
The approach taken in this commit is to remove input type checking from
the dispatching logic. Prior to this commit, the `prototypeMethod`
function would return `null` for Arrays or Objects that don't have
Setoids on each of their values. But the only code path benefitting from
this extra check were the Setoid and Ord `.test` methods.

This commit separates the extra checks out into their own function,
called `hasPrototypeMethod`, to be used exclusively by the testing
methods. The dispatching logic now assumes that input was already
tested, and so behaves the same for sane input, but possibly differently
for garbage input.

This leads to a very significant performance increase for equals or lte
dispatches on nested structures.