Simplify code structure
Make it so `TypeClass.methods.<method>` functions take the context (the
eventual `this` value) as a last argument instead of a first argument.
Taking the `this` value last is the most common structure of Fantasy
Land dispatchers, making it so most of the `Z.<function>` exports are
now directly equal to the corresponding dispatcher function on the
TypeClass.

As a result, it also becomes more obvious which exported functions
explicitly flip the arguments before passing them to the dispatchers.